pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai10101ia"
version = "0.1.0"
description = "Fondation GitHub contrôlée par checklist maître et registre machine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0.1,<7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
